+++ src/config.rs
@@ -6,6 +6,7 @@ use crate::errors;
 
 use std::fs::File;
 use std::io::Read;
+use std::path;
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 #[serde(untagged)]
@@ -53,10 +54,10 @@ pub(crate) struct FlokiConfig {
 }
 
 impl FlokiConfig {
-    pub fn from_file(file: &str) -> Result<FlokiConfig, Error> {
+    pub fn from_file(file: &path::Path) -> Result<FlokiConfig, Error> {
         let mut f = File::open(file).map_err(|e| {
             errors::FlokiError::ProblemOpeningConfigYaml {
-                name: file.to_string(),
+                name: file.display().to_string(),
                 error: e,
             }
         })?;
@@ -64,13 +65,13 @@ impl FlokiConfig {
         let mut raw = String::new();
         f.read_to_string(&mut raw)
             .map_err(|e| errors::FlokiError::ProblemReadingConfigYaml {
-                name: file.to_string(),
+                name: file.display().to_string(),
                 error: e,
             })?;
 
         let config =
             serde_yaml::from_str(&raw).map_err(|e| errors::FlokiError::ProblemParsingConfigYaml {
-                name: file.to_string(),
+                name: file.display().to_string(),
                 error: e,
             })?;