+++ CHANGELOG.md
@@ -8,6 +8,8 @@ version number is tracked in the file `VERSION`.
 
 ## Unreleased
 ### Changed
+- Rename `FLOKI_HOST_WORKDIR` to `FLOKI_HOST_MOUNTDIR` - BREAKING
+- Also search ancestors of the working directory for a `floki.yaml` - MINOR
 ### Added
 
 ## [0.2.0] - 2019-08-10